requirements.txt
langchain==0.2.11
openai==1.37.0
tabulate==0.9.0
python-environ==0.4.54
ffmpeg==1.4
tiktoken==0.7.0
lark==1.1.9
streamlit==1.36.0
pydantic==1.10.8 # https://github.com/hwchase17/langchain/issues/7548
python-dotenv==1.0.1
langchain_openai==0.1.17
langchain_community==0.2.10

# Vector DB
pinecone-client==5.0.0
chromadb==0.5.3
langchain-chroma==0.1.2

# Confluence
atlassian-python-api==3.41.14
lxml==5.2.2
markdownify==0.13.1
